ci(e2e): split desktop jobs and isolate renderer layout checks
