+++ e2e/fixtures/localized-settings.ts
@@ -1,0 +1,154 @@
+export const localizedSettingsCases = [
+  {
+    language: 'Simplified Chinese',
+    pickerLabel: '简体中文',
+    locale: 'zh-Hans',
+    model: '模型',
+    projects: '项目',
+    modelSettings: '模型设置',
+    settings: '设置',
+    openNavigation: '打开设置导航',
+    general: '通用',
+    appearance: '外观',
+    interfaceLanguage: '界面语言',
+    mainModel: '主模型',
+    scenarioModels: '场景模型',
+    expandSubagent: '展开子智能体设置',
+    reasoningEffort: '推理强度',
+    defaultEffort: '默认',
+    closeSettings: '关闭设置'
+  },
+  {
+    language: 'Traditional Chinese',
+    pickerLabel: '繁體中文',
+    locale: 'zh-Hant',
+    model: '模型',
+    projects: '專案',
+    modelSettings: '模型設定',
+    settings: '設定',
+    openNavigation: '開啟設定導覽',
+    general: '一般',
+    appearance: '外觀',
+    interfaceLanguage: '介面語言',
+    mainModel: '主模型',
+    scenarioModels: '情境模型',
+    expandSubagent: '展開子智能體設定',
+    reasoningEffort: '推理強度',
+    defaultEffort: '預設',
+    closeSettings: '關閉設定'
+  },
+  {
+    language: 'Japanese',
+    pickerLabel: '日本語',
+    locale: 'ja',
+    model: 'モデル',
+    projects: 'プロジェクト',
+    modelSettings: 'モデル設定',
+    settings: '設定',
+    openNavigation: '設定ナビゲーションを開く',
+    general: '一般',
+    appearance: '外観',
+    interfaceLanguage: '表示言語',
+    mainModel: 'メインモデル',
+    scenarioModels: 'シナリオモデル',
+    expandSubagent: 'サブエージェント設定を展開',
+    reasoningEffort: '推論強度',
+    defaultEffort: 'デフォルト',
+    closeSettings: '設定を閉じる'
+  },
+  {
+    language: 'Korean',
+    pickerLabel: '한국어',
+    locale: 'ko',
+    model: '모델',
+    projects: '프로젝트',
+    modelSettings: '모델 설정',
+    settings: '설정',
+    openNavigation: '설정 탐색 열기',
+    general: '일반',
+    appearance: '외관',
+    interfaceLanguage: '인터페이스 언어',
+    mainModel: '메인 모델',
+    scenarioModels: '시나리오 모델',
+    expandSubagent: '서브에이전트 설정 펼치기',
+    reasoningEffort: '추론 강도',
+    defaultEffort: '기본값',
+    closeSettings: '설정 닫기'
+  },
+  {
+    language: 'Russian',
+    pickerLabel: 'Русский',
+    locale: 'ru',
+    model: 'Модель',
+    projects: 'Проекты',
+    modelSettings: 'Настройки модели',
+    settings: 'Настройки',
+    openNavigation: 'Открыть навигацию по настройкам',
+    general: 'Общие',
+    appearance: 'Внешний вид',
+    interfaceLanguage: 'Язык интерфейса',
+    mainModel: 'Основная модель',
+    scenarioModels: 'Сценарные модели',
+    expandSubagent: 'Развернуть настройки: Субагент',
+    reasoningEffort: 'Глубина рассуждений',
+    defaultEffort: 'По умолчанию',
+    closeSettings: 'Закрыть настройки'
+  },
+  {
+    language: 'French',
+    pickerLabel: 'Français',
+    locale: 'fr',
+    model: 'Modèle',
+    projects: 'Projets',
+    modelSettings: 'Paramètres du modèle',
+    settings: 'Paramètres',
+    openNavigation: 'Ouvrir la navigation des paramètres',
+    general: 'Général',
+    appearance: 'Apparence',
+    interfaceLanguage: "Langue de l'interface",
+    mainModel: 'Modèle principal',
+    scenarioModels: 'Modèles de scénario',
+    expandSubagent: 'Développer les paramètres de Sous-agent',
+    reasoningEffort: 'Effort de raisonnement',
+    defaultEffort: 'Par défaut',
+    closeSettings: 'Fermer les paramètres'
+  },
+  {
+    language: 'Spanish',
+    pickerLabel: 'Español',
+    locale: 'es',
+    model: 'Modelo',
+    projects: 'Proyectos',
+    modelSettings: 'Configuración del modelo',
+    settings: 'Configuración',
+    openNavigation: 'Abrir navegación de configuración',
+    general: 'General',
+    appearance: 'Apariencia',
+    interfaceLanguage: 'Idioma de la interfaz',
+    mainModel: 'Modelo principal',
+    scenarioModels: 'Modelos de escenario',
+    expandSubagent: 'Expandir configuración de Subagente',
+    reasoningEffort: 'Esfuerzo de razonamiento',
+    defaultEffort: 'Predeterminado',
+    closeSettings: 'Cerrar configuración'
+  },
+  {
+    language: 'German',
+    pickerLabel: 'Deutsch',
+    locale: 'de',
+    model: 'Modell',
+    projects: 'Projekte',
+    modelSettings: 'Modelleinstellungen',
+    settings: 'Einstellungen',
+    openNavigation: 'Einstellungsnavigation öffnen',
+    general: 'Allgemein',
+    appearance: 'Darstellung',
+    interfaceLanguage: 'Sprache der Benutzeroberfläche',
+    mainModel: 'Hauptmodell',
+    scenarioModels: 'Szenariomodelle',
+    expandSubagent: 'Unteragent-Einstellungen erweitern',
+    reasoningEffort: 'Reasoning-Aufwand',
+    defaultEffort: 'Standard',
+    closeSettings: 'Einstellungen schließen'
+  }
+] as const
