+++ tsconfig.web.json
@@ -6,12 +6,18 @@
     "src/renderer/src/**/*.tsx",
     "src/renderer/web/**/*",
     "src/preload/*.d.ts",
-    "src/shared/**/*"
+    "src/shared/**/*",
+    "e2e/browser/fixture/**/*"
   ],
   "compilerOptions": {
     "composite": true,
     "noImplicitAny": true,
-    "lib": ["ESNext", "DOM", "DOM.Iterable", "DOM.AsyncIterable"],
+    "lib": [
+      "ESNext",
+      "DOM",
+      "DOM.Iterable",
+      "DOM.AsyncIterable"
+    ],
     "jsx": "react-jsx",
     "baseUrl": ".",
     "paths": {
